+++ pyreports/core.py
@@ -24,6 +24,12 @@
 
 # region Imports
 import tablib
+import smtplib
+import ssl
+from email.mime.text import MIMEText
+from email.mime.multipart import MIMEMultipart
+from email import encoders
+from email.mime.base import MIMEBase
 from .io import FileManager
 from .exception import ReportManagerError, ReportDataError
 
@@ -245,7 +251,7 @@ class Report:
 
         :return: string
         """
-        return self._print_data()
+        return str(self._print_data())
 
     def _print_data(self):
         """Print data and count
@@ -294,6 +300,64 @@ class Report:
                 print(self)
         else:
             raise ReportManagerError('the output object must be FileManager or NoneType object')
+
+    def send(self, server, _from, to, cc=None, bcc=None, subject=None, body='', auth=None, _ssl=True, headers=None):
+        """Send saved report to email
+
+        :param server: server SMTP
+        :param _from: email address 'from:'
+        :param to: email address 'to:'
+        :param cc: email address 'cc:'
+        :param bcc: email address 'bcc:'
+        :param subject: email subject. Default is report title
+        :param body: email body
+        :param auth: authorization tuple "(user, password)"
+        :param _ssl: boolean, if True port is 465 else 25
+        :param headers: more header value "(header_name, key, value)"
+        :return: None
+        """
+        if not self.output:
+            raise ReportDataError('if you want send a mail with a report in attachment, must be specified output')
+
+        # Prepare mail header
+        message = MIMEMultipart("alternative")
+        message["Subject"] = self.title if not subject else subject
+        message["From"] = _from
+        message["To"] = to
+        if cc:
+            message["Cc"] = cc
+        if bcc:
+            message["Bcc"] = bcc
+        if headers:
+            message.add_header(*headers)
+
+        # Prepare body
+        part = MIMEText(body, "html")
+        message.attach(part)
+
+        # Prepare attachment
+        self.export()
+        attach_file_name = self.output.file
+        attach_file = open(attach_file_name, 'rb')
+        payload = MIMEBase('application', 'octate-stream')
+        payload.set_payload(attach_file.read())
+        encoders.encode_base64(payload)
+        payload.add_header('Content-Disposition', 'attachment', filename=attach_file_name)
+        message.attach(payload)
+
+        # Prepare SMTP connection
+        if _ssl:
+            port = smtplib.SMTP_SSL_PORT
+            protocol = smtplib.SMTP_SSL
+            kwargs = {'context': ssl.create_default_context()}
+        else:
+            port = smtplib.SMTP_PORT
+            protocol = smtplib.SMTP
+            kwargs = {}
+        with protocol(server, port, **kwargs) as srv:
+            if auth:
+                srv.login(*auth)
+            srv.sendmail(_from, to, message.as_string())
 
 
 class ReportBook:
@@ -394,5 +458,24 @@ class ReportBook:
         else:
             for report in self:
                 report.export()
+
+    def send(self, server, _from, to, cc=None, bcc=None, subject=None, body='', auth=None, _ssl=True, headers=None):
+        """Send saved report to email
+
+        :param server: server SMTP
+        :param _from: email address 'from:'
+        :param to: email address 'to:'
+        :param cc: email address 'cc:'
+        :param bcc: email address 'bcc:'
+        :param subject: email subject. Default is report title
+        :param body: email body
+        :param auth: authorization tuple "(user, password)"
+        :param _ssl: boolean, if True port is 465 else 25
+        :param headers: more header value "(header_name, key, value)"
+        :return: None
+        """
+        for report in self:
+            report.send(server, _from, to, cc=cc, bcc=bcc, subject=subject, body=body, auth=auth,
+                        _ssl=_ssl, headers=headers)
 
 # endregion